Modularize code and define library members
- The code has been edited to comply with SOLID principles as much as possible.
- Library members were defined and the books they borrowed and the time they borrowed were recorded.
- A method has been added to find records of books whose borrowing period has expired.

+++ Classes/Book.cs
@@ -8,7 +8,7 @@ namespace LibraryApp.Classes
 {
     internal class Book
     {
-        public int ISBSN { get; set; }
+        public ulong ISBSN { get; set; }
         public string Title { get; set; }
         public string Author { get; set; }
         public int NumberOfCopies { get; set; }
